+++ lib/cta/api.rb
@@ -3,7 +3,6 @@
 require_relative "api/version"
 require_relative "api/errors"
 require_relative "api/response"
-require_relative "api/shared"
 require_relative "api/bus_tracker"
 require_relative "api/train_tracker"
 require_relative "api/customer_alerts"